Fixed functions.
Added default commands for elevator, intake pivot, and shooter pivot that are based on joystick input.

Deleted swerve code in anticipation of using CTRE swerve.

Formatted both pivot subsystems to be more in line with the way we will code them.

Added a timer to constants and added more commands to test the functions of the robot.